# docs/openapi/all-services.yaml
openapi: 3.1.0
info:
  title: BeckLab Tools — Full API Reference
  version: 1.0.0
  description: Complete API reference for all BeckLab micro-services. Each service is independently accessible at its own subdomain.
  contact:
    name: BeckLab
    url: https://tools.becklab.cloud
servers:
  - url: https://hash.tools.becklab.cloud
    description: Hash Service
  - url: https://short.tools.becklab.cloud
    description: URL Shortener
  - url: https://convert.tools.becklab.cloud
    description: Base64/Converter
  - url: https://cron.tools.becklab.cloud
    description: Cron Jobs
  - url: https://dns.tools.becklab.cloud
    description: DNS Monitor
  - url: https://webhook.tools.becklab.cloud
    description: Webhook Relay
  - url: https://fmt.tools.becklab.cloud
    description: YAML/JSON Tool
  - url: https://qr.tools.becklab.cloud
    description: QR Generator
  - url: https://editor.tools.becklab.cloud
    description: Image Editor
  - url: https://dashboard.tools.becklab.cloud
    description: BeckFlow Gateway
  - url: https://auth.tools.becklab.cloud
    description: Auth Service
...
paths:
...
  /api/v1/dns:
    post:
      summary: Perform DNS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, example: "example.com" }
                recordType:
                  type: string
                  enum: [A, AAAA, MX, TXT, CNAME, NS, SOA]
                  default: A
      responses:
        '200': { description: DNS lookup results }
...
  /api/v1/qr:
    post:
      summary: Generate QR code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data: { type: string }
                format: { type: string, enum: [png, svg], default: png }
                size: { type: integer, default: 256 }
                errorCorrection: { type: string, enum: [L, M, Q, H], default: M }
                template: { type: string, enum: [none, wifi, vcard, email, sms, geo] }
      responses:
        '200': { description: QR code image }
  # === Image Editor API ===
...
  /api/v1/workflow:
    post:
      summary: Execute a BeckFlow workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [steps]
              properties:
                steps:
                  type: array
                  items:
                    type: object
                    required: [service, endpoint, method, body]
                    properties:
                      service: { type: string }
                      endpoint: { type: string }
                      method: { type: string }
                      body: { type: object }
                      usePrevious: { type: boolean }
      responses:
        '200': { description: Workflow execution results }
...